Add new functionality to datalogger examples. Geodetic WGS84 columns latitude, longitude, direction, altitude now appear in all machine specific and generic examples regardless of whether original coordinates sent from the client were localised. Transform revision and the transform source file are now also output. Include names for RDM objects such as user and delay which were previously only tracked by identifier.
This revision demonstrates the established process for deriving the L,L,H from localised position data provided by datalogger. Doing so requires interrogating the site for its localisation history so that the correct approximation matrices can be used appropriate for the localisation that was active at the site when the data was collected. Because localisations can be updated, code can’t simply apply the current one to all historical data.

1.	For each datalogger Replicate providing positions in local coordinates, record the transform revision in use at the time of the replicate and cache that information.
2.	Batch process the local to WGS84 conversion via the transform service at the end of processing for each transform revision active over the time range that datalogger data was provided (too costly to convert a point at a time).
3.	Output the L,L,H to the appropriate columns as would have been natively done for replicates from a haul app.
4.	Also populate a new column to the CSV output that shows the transform revision used for this conversion and the source file so it’s clear to the user what was used to do the conversion and whether or not that changed over the time span.

The end result is a CSV output that can be loaded in Excel. An Excel map add-in to plot the lat, long columns can be used to visualise these columns.

Another important consideration is which local point the user intends to transform for each machine. Because the MFK standard supports multiple points of interest on a machine, there are potentially multiple local points that can be converted. This is obviously machine specific. A dozer and grader for example will report left and right of blade in local coordinates whereas an excavator will report bucket points. Multiple articulation points on the machine body are also available.

This commit uses the “local_position” point as a proxy for the machine position so that a representative position for each machine is converted to L,L,H. Multi component machines contribute only one “local_position” point from Replicate messages to avoid corrupting the single row per replicate paradigm in the CSV output. Otherwise multiple component machines (excavators) would add two local points for every one row added to the CSV output, adding an incorrect row offsets to the WGS84 columns when appended on final output.

To achieve this, this commit introduces a major refactor of similar datalogger to CSV code used in machine specific and generic examples into the one function ProcessDataloggerToCsv. Introduce GeodeticCoordinateManager and TransformManager to manage all WGSS84 output at the final file write.

Refactor the query and download of localization history into query_and_download_site_localization_file_history. Refactored the creation of the object that gets written to the CSV output into wgs84_coord_to_object_list_item and accommodated the two different height keys, alt & height.

Added helper function get_transform_revision_for_time to return the revision of the transform object that was in effect at a specified time at a site.